added the 'request' parameter to BaseTable.__init__
this allows views to simply pass along the HttpRequest object to set
the sort order, rather than extracting it manually. the name of the
GET parameter used can be specified by the 'sort_param' argument or
option.

this works for both BaseTables and ModelTables, turning:

    MyTable(order_by=req.GET.get('sort', None))

into:

    MyTable(request=req)

which i think is much nicer. it would be even nicer if the pagination
stuff could be extracted, too...